fix: 모임 생성 및 수정 관련 DTO 수정
