+++ Inside-Out/src/main/java/com/goorm/insideout/meeting/dto/request/MeetingUpdateRequest.java
@@ -11,9 +11,7 @@ import com.goorm.insideout.meeting.domain.Meeting;
 import com.goorm.insideout.meeting.domain.MeetingPlace;
 import com.goorm.insideout.user.domain.User;
 
-import lombok.AccessLevel;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 @Getter
 public class MeetingUpdateRequest {
@@ -25,7 +23,7 @@ public class MeetingUpdateRequest {
 
 	private String categoryDetail;
 
-	private MeetingCreateRequest.MeetingPlaceRequest meetingPlace;
+	private MeetingPlaceRequest meetingPlace;
 
 	private int participantLimit;
 
@@ -75,35 +73,5 @@ public class MeetingUpdateRequest {
 			meetingPlace,
 			rules
 		);
-	}
-
-	@Getter
-	@NoArgsConstructor(access = AccessLevel.PRIVATE)
-	private static class MeetingPlaceRequest {
-		private String name;
-
-		private String placeUrl;
-
-		private Long kakaoMapId;
-
-		private String addressName;
-
-		private String roadAddressName;
-
-		private Double latitude;
-
-		private Double longitude;
-
-		private MeetingPlace toEntity() {
-			return MeetingPlace.createMeetingPlace(
-				name,
-				placeUrl,
-				kakaoMapId,
-				addressName,
-				roadAddressName,
-				latitude,
-				longitude
-			);
-		}
 	}
 }
